moves the 3 libs used from magnustools to onboard
This is a terrible idea.  It's something I'm exploring to attempt to
understand what the right solution should be.  The files used from
magnustools appear to be self contained libraries.  I've picked them up
and dropped them in a libs folder on the quicktools repo.  Assuming I'm
right, this should work fine.  The files in question are post
modificaitons to make it work at Matrix.

The next step is generalizing those libraries to understand the
sites.json file.