WikidataItemList takes a site at construction
The WikidataItemsList class has been modified to require a site as
described in the sites.json file at construction time to determine
which wiki it will talk to.  The api url's are constructed from that
structure and then used throughout the call stack.

The calls to it have been updated as well.

Also, formatting was updated.